+++ src/api/entities/TempAttachment.ts
@@ -18,8 +18,8 @@ export class TempAttachment implements TempAttachmentModel {
 
   @Column({ type: 'text' })
   ext: string;
-  
-  @Column({ type: 'int' })
+
+  @Column({ type: 'bigint' })
   size: number;
 
   @Column({ type: 'text', nullable: true })
